Move FindModuleCache and BuildSource into new file modulefinder.py
